+++ app/models/automated_checks/dqt_report_csv_to_records.rb
@@ -12,6 +12,7 @@ module AutomatedChecks
         {
           claim_reference: grouped_rows.first.fetch(CLAIM_REFERENCE_COLUMN),
           qts_date: parse_date(grouped_rows.first.fetch("dfeta qtsdate")),
+          surname: extract_surname(grouped_rows.first.fetch("fullname")),
           date_of_birth: parse_date(grouped_rows.first.fetch("birthdate")),
           degree_jac_codes: collate_fields_from_rows(%w[HESubject1Value HESubject2Value HESubject3Value], grouped_rows),
           itt_subject_jac_codes: collate_fields_from_rows(%w[ITTSub1Value ITTSub2Value ITTSub3Value], grouped_rows)
@@ -29,6 +30,10 @@ module AutomatedChecks
       Date.parse(date)
     rescue
       nil
+    end
+
+    def extract_surname(fullname)
+      (fullname || "").split(" ").last
     end
   end
 end